remote/app/function_process/requirements.txt
python-valve==0.2.1
cfnresponse==1.1.1
crhelper>=2.0.10,<3.0.0